+++ inventory/frontend/app/components/Icon.vue
@@ -1,3 +1,4 @@
+/
 <script setup lang="ts">
 const props = defineProps<{ name: string; size?: string }>();
 const sz = computed(() => props.size ?? "18px");
