+++ inventory/frontend/utils/types/inventory.ts
@@ -1,0 +1,7 @@
+export interface InventoryItem {
+  id: number;
+  name: string;
+  quantity: number;
+  category: string;
+  barcode: string;
+}
